Cataracts Preprocs added

+++ data/data.py
@@ -76,8 +76,8 @@ class SEGDataset(Dataset):
 
         if self.transform is not None:
             transformer = self.transform(image = image, mask = mask)
-            image, mask = transformer["image"], transformer["mask"]
-        
+            image, mask = transformer["image"], transformer["mask"].long()
+
         return image, mask
